Create README.md
# City car simulator

The program allows you to simulate driving around the city. Finally, graphs are generated in gnuplot.
# Licence

Creator / authors allow anyone to view any file in this project. The use of any file or fragment thereof in ant projects, publicly sharing file or a fragment thereof, as well as other activities without the permit of pawel112 is prohibited.